🔨 Created context folder for contexts
