+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import { AppProps } from 'next/app';
 import { ThemeProvider } from 'styled-components';
 
-import FavoritesProvider from '../hooks/useFavorites';
+import FavoritesProvider from '../context/favoriteContext';
 
 import GlobalStyles from '../styles/global';
 import theme from '../styles/theme';
